Danny and Grace, please review.  This checkin addresses your comments on checkins 14 and 15.
Checkin 14:

>TraversalMethod.java:
>Do you think we should rename this to Traverser?

Done in this checkin.

>ConnectorConfigStore.java:
>We also need a way to access the schedule.

I agree.  We need an object that holds a schedule.  Could you and Phuong agree on the semantics?  Then could you please add it?

Checkin 15:

>Pusher.java:
>Should we have something with more typing?  If not, do we need to
>specify what needs to go into the PropertyMap?

There is documentation of this kind in spi/QueryTraversalManager.java

>Manager.java:
>We need to consider whether we really want Document objects or whether
>we want configuration objects passed around.

I agree.  Let's talk about this.

>107:  why is this returning a ResultSet?  I thought ResultSet is just
>for queries.

Cut & paste error!  Fixed in this checkin.