Put it all together!
This checkin contains glue code.  A new class, "Context", provides access to all the important singletons in the system (through Spring) and starts up processing.  It has a unit test - which mainly makes sure that everything is there.  The servlets have been changed to get the manager through the Context.  The Context is aware of whether we're testing or in production, and instantiates the right kind of Spring applicationContext.  Thus the MockManager is now moved to where it belongs, the javatests hierarchy.  A new Manager implementation "ProductionManager" is added as the non-mock manager.

Two applicationContext.xml config files are provided for now.  The default lives in etc/applicationContext.xml, which is mapped to WEB-INF/applicationContext.xml when the war file is built.  At present, this config file creates a real instantiator and scheduler, but still uses a MockUrlConn, MockConnectorStateStore.  It also uses a HashMapMonitor.  Most of the lines of the file go to instantiating a schedules list - I'm sure we'll have a better way to do that soon.  

The second context is testdata/servletMockContext/applicationContext.xml.  It instantiates the mockmanager and nothing else.  To run tests of the servlets in tomcat againt the mockmanager, install the war then replace WEB-INF/applicationContext.xml with this one.

Danny and Meghna, please both take a look at this, especially at the applicationContext.xml.  We will be making changes to this file as we check in new production components.