WorkQueue.java: Add way to cancel a work item in the WorkQueue.
TraversalScheduler.java:
Add method to call into TraversalScheduler when removing a connector.  This method will interrupt the thread 
if it is running or just remove it from the WorkQueue if it isn't running.  When a connector is added, the 
TraversalScheduler will automatically pick it up the next time it reads the schedules.

Also rename TraversalRunnable to TraverserWorkQueueItem.

TraversalSchedulerTest.java:
Added unittest for cancelWork method.


Max, please review.