Implement setManagerConfig.  Danny, please review.
Architectuyral change: to applicationContext: we now use a properties file for the stuff that can be controlled through setManagerConfig.  The properties file should be pointed to in the applicationContext itself, by a special bean for that purpose: "ApplicationContextProperties".  The Context looks for this bean to find the name of the properties file to update.  After updating, the Context restarts pretty much everything it can get its hands on.

applicationContext.xml: two new beans: ApplicationContextProperties and a PropertyPlaceholderConfigurer to do the actual properties injection.  See the change to the FeedConnection bean.

InstanceInfo.java: refactor two methods to make public so Context can see them.

Context.java: core implementation of setManagerConfig

ProductionManager.java: wire new implementation in

MockPusher.java: add default constructor (for the new test applicationContext.xml)

New files:

etc/applicationContext.xml - this needs to be packed in to the war
SetManagerConfigTest.java - new test
testdata/contextTests/** - new data files for this test