The proposed fix to Bug 895651.  Added support to allow the setting of the action attribute on feeds by adding a property to the document.
The DocPusher will now check the document for an optional
SpiConstants.PROPNAME_ACTION property.  Current legal values are:

  SpiConstants.ActionType.ADD.toString();
  SpiConstants.ActionType.DELETE.toString();

The behaviors are outlined below.

1) If the 'action' property is not present, the feed will not contain an
'action' attribute and the feeder defaults to the 'add' action.  This is the
same as the current behavior.

2) If the 'action' property is specified and it it a legal
SpiConstants.ActionType then it will be set as the action attribute on the feed.
Current legal actions are 'add' and 'delete'.

3) If the 'action' property is specified and is NOT a legal
SpiConstants.ActionType, a WARNING will be logged and it will be ignored.  NOTE:
This will result in default behavior of the feeder using the 'add' action.

Let me know if behavior #3 is problem.  Alternate behavior is to log an ERROR
and drop the document.