This change addresses customer problem #395342092
A customer encountered a race condition between the
TraversalScheduler thread and TraversalWorkItem threads
it waits on.  The following is an analysis excerpted from
the email thread discussing the problem:

I've been looking into it.  It is not a hanging work thread.
The connector is cranking along.  Each call to runBatch() returns
only 10 documents, but it does so quite quickly, so it traverses
100 documents in less than a minute.  It then waits out the rest
of the minute spinning - checking once per second and adding 20
lines of log entries, the only meaningful one being
    "remainingDocsToTraverse=0"

The hostLoadManager is waiting for the minute to expire before
returning a non-zero batchHint().  For some reason, the check
for non-zero batchHint is done inside the workItem thread rather
than the scheduler thread.  This forces the workItem to get
scheduled once per second, only to say "I got nothin".  This 
check should probably be moved to HostLoadManager.shouldDelay().

But the once-per-second spamming actually points to the failure.
While waiting for the minute to expire, each second we usually see:

key: [S] Scheduler Thread    [W] Worker Thread

[S]: Trying to add traversal work to workQueue
[S]: Adding work: TraversalWorkQueueItem
[S]: Beginning wait (timeout=MAX_INT)...
 - Scheduler Thread waits for signal from Worker
[W]: Removing work: TraversalWorkQueueItem
[W]: maxDocsPerPeriod=100, docsTraversed=100, remainingDocsToTraverse=0
 - Worker thread realizes that it can do no work
 - Worker thread finishes workItem and signals Scheduler
[S]: ...ending wait  (after receiving signal)

However in the case of the big gap, we see this:

[S]: Trying to add traversal work to workQueue
[S]: Adding work: TraversalWorkQueueItem
[W]: Removing work: TraversalWorkQueueItem
[W]: maxDocsPerPeriod=100, docsTraversed=100, remainingDocsToTraverse=0
 - Worker thread realizes that it can do no work
 - Worker thread finishes workItem and signals Scheduler
[S]: Beginning wait (timeout=MAX_INT)...
 - Scheduler Thread waits for signal from Worker
 - 8.75 hour delay
 - Servlet thread logging HTTP traffic

Basically, once made runnable, the Worker thread ran to
completion before the Scheduler thread started to wait().
The Worker sent the notifyAll() before the Scheduler
called wait(), so the notification appears to have been
missed.

This could have happened even if the Sharepoint Connector
did not supply a huge timeout, as a timeout of 0
(wait forever) would be used instead to the same effect
(plus or minus a couple hundred years).


Change Log:
----------
M  projects/connector-manager/source/java/com/google/enterprise/connector/scheduler/TraversalScheduler.java
   - Move the access to isFinished booleans to inside
     synchronized blocks.

M  projects/connector-manager/source/java/com/google/enterprise/connector/scheduler/HostLoadManager.java
   - Enhance shouldDelay() method to include the check
     for numDocsTraversed exceeding maxDocsPerPeriod.
     This avoids scheduling the WorkItem at all if it
     is waiting for its traversal period to expire.