Description:
Fix for Issue 146.  Added an index.html page to each of the WAR files that
redirects to the TestConnectivity Servlet.  This way, when the Web Application
is deployed on a server the customer can navigate to the root of the Web
Application (http://<host>:<port>/connector-manager) and get some status
information about the server rather than a 404.

Tests Passed:

  - All Unit tests.
  - Tested without redirect and link on index page will navigate to the
    TestConnectivity Servlet.

ChangeLog:

A source/html/index.html
  - New Home Page for Web Application that redirects to the TestConnectivity
    Servlet.

M build.xml
  - Added property for web content location and added FileSet to each of the WAR
    targets.