Fix Issue 232: ThreadPool Memory Leak
The Connector Manager ConectorCoordinator submits traversal batch jobs to the
ThreadPool to execute.  Traversal batches are cancelable and may time out,
so ThreadPool wraps the CancelableBatch with a CancelTimeoutRunnable,
TimeoutTask and some FutureTasks.  It builds up a dependency graph
(with a cycle) of the following objects:
  * QueryTraverser
  * BatchSize
  * BatchCoordinator
  * CancelableBatch
  * TimeoutTask
  * CancelTimeoutRunnable
  * TaskHandle
  * FutureTask (x 2, as taskFuture and timeoutFuture)

The FutureTasks are then submitted to the CompletionService and the
TimeoutService.  These (along with a few other objects internal to
CompletionService and TimeoutService), are what are getting leaked.

The expected behaviour is that once the traversal batch run finishes,
the CompletionTask will pick up the results from taskFuture, and cancel
the timeoutFuture.  If the traversal takes too long, the scheduled
timeoutFuture task runs and call TimeoutTask.timeout() which cancels the
batch and eventually cancels its taskFuture.  Cancelling and/or
competing the two FutureTasks is supposed to break this collection of
objects off from active ownership so that they can get garbage collected.

It took more than a bit of analysis to find out where things go wrong.
I then realized that completionService.submit() takes a Runnable
(or Callable) and *returns* a Future (actually another instance
of FutureTask).  Similarly, timeoutService.schedule() takes a Runnable
(or Callable) and returns a Future.  The FutureTasks created by
ThreadPool are not the Futures that the completionService and
timeoutService are holding.  Those returned Futures were discarded
by ThreadPool.

So when the traversal batch completes, the completionService
reaps the real Future for the CancelTimeoutRunnable (leaving
the bogus taskFuture hanging about), then cancels the bogus
timeoutFuture (leaving the real timoutTask future languishing
in the timeoutService).  Since the timeoutService still holds
the real timeout Future, the whole dependency graph of objects
hangs out.  Eventually the timeout service will finally run the
scheduled timeout task (after an hour), which will attempt to
cancel the long-finished traversal batch (harmlessly).  At this
point the objects might finally be disconnected enough to be
garbage collected.

The good news is the objects eventually will get reaped, even
if it may take an hour or two.  The bad news is several connectors
(including diffing connectors like the FileSystem connector),
share a subtly bad behavior.  Diffing connectors will return
a non-null, but empty DocumentList if their ChangeQueue runs
dry. This is a coordinated signal to the Connector Manager to
immediately call resumeTraversal, rather than wait for the
RetryDelay interval.  This coordinated signal was designed to
help connectors process long-running jobs without timing out.
(Some complex SQL queries might run for 30 or 40 minutes.)
The end result is that the Connector Manager might call startBatch
every second, for minutes on end, waiting for the connector to
return a non-empty DocumentList.  This behaviour is what causes
many thousands of these object clusters to pile up, unreaped.

This change dispenses with the bogus FutureTasks that wrap
TimeoutTask and CancelTimeoutRunnable, and populates timeoutFuture
and taskFuture with the *real* futures returned by the timeoutService
and completionService.  This required a small amount of re-organization
of the code because of the cycle of dependecies built up:

TaskHandle --> taskFuture --> CancelTimeoutRunnable
^                                    |
|-- TimoutTask <-- timeoutFuture <---|

See:
http://code.google.com/p/google-enterprise-connector-manager/issues/detail?id=232

Change Log:
----------
M projects/connector-manager/source/java/com/google/enterprise/connector/instantiator/ThreadPool.java

Code Review:
-----------
http://codereview.appspot.com/5675063