Sort metadata in feeds so GSA detects non-updates
The GSA performs a string comparison on the metadata to see if it
changed from the metadata it has indexed. This check will fail if the
same metadata is presented in a different order. Thus, this change
writes metadata to feeds in a canonical form so that the random behavior
of HashSets or the like does not cause a performance degradation.

Code review: http://codereview.appspot.com/5969075/