Require use of HTTP Basic for private Documents
If the GSA is not informed that a document is private at crawl time,
then it will be marked public, even if a feed file marked it as private.
As a signal to the GSA, this requires private documents (ones in which
ACLs can not be represented) be accessed only after the GSA has used
HTTP Basic.

The actual username/password combination doesn't matter, just the fact
that the GSA was required to specify them. Thus, we only check to see
if the GSA provided the Authorization header and don't actually check
the credentials.

Code review: http://codereview.appspot.com/5979047/