Extracted TraversalTimer and EmptyDocumentList classes from the Documentum connector for us in the Livelink and SharePoint connectors.
The connectors have loops that continue to do work looking for
documents to return. So that these loops do not lead to the batch
being cancelled, and so that they can preserve their work, the loops
contain a timer based on TraversalContext#traversalTimeLimitSeconds,
and if half the time limit has passed, an empty document list with a
new checkpoint is returned. The empty document list ensures that the
progress made by the loop is saved (compared to returning null).

Code review: http://codereview.appspot.com/2796041/