Connector release repackaging: Connector Manager
Adds a repackage target to rename the downloads from
connector-manager-${svn.branch} to connector-manager-${version}.
This is intended for repackaging release candidates for release,
but it will work from any branch.

Removes the generation of .tar.gz downloads.

Fixes a small issue in svnbuild when repackaging source archives.

Code Review:
http://codereview.appspot.com/9312044