Fix test failures uncovered by the removal of the document store.
The failures were errors in the test gear and expected results. The
removal of the document store merely uncovered one of the incorrect
results. In all cases, an unexpected exception was thrown that dropped
the entire batch on the floor, instead of just skipping a document as
expected. The tests expected 0 documents in the batch count, but the
actual count was higher (either 1 or 2, depending on whether the
documents were skipped in nextDocument or take).

Added assertions for the number of documents pushed, which differs
from the batch count, which includes documents that were skipped in
Pusher.take.

Delayed some assertions in ValidatingPusher, which were throwing
Errors that dropped the batch on the floor, to turn them into test
failures instead. The assertion failures were happening because the
call to skipDocument was in the wrong place, in throwException. The
internal counter in ValidatingPusher handles skipped documents in
take, but doesn't know about skipped documents in nextDocument, so
those are the ones where skipDocument needs to be called.

Added handling for RuntimeException in ValidatingPusher.take. The
existing tests don't require this, but the code was fragile, and could
miscount the expectedId value in ValidatingPusher if a
RuntimeException were thrown before the last document in a batch.

Other changes:

* Cleaned up QueryTraverser a little from the removal of the document store.
* Uncommented @Override annotations.

Code review: https://codereview.appspot.com/12936045/