Merge changes to ConnectorFields since it was copied from the LDAP connector.
The LDAP connector was not modified to use the ConnectorFields class
after it was copied to the CM util package in r2616, and several
updates have been made to the LDAP copy in the meantime.

This change minimizes the differences between the two copies, and adds
generic support for the new LDAP features. The LDAP connector added
two attributes to the checkbox inputs (disabled and onclick), and
wrapped the inputs in a label. I've added the label and a callback
interface that allows users of the ConnectorFields class to add such
attributes without needing to know about them in ConnectorFields
itself.

Code review: https://codereview.appspot.com/41700043