Eviscerate SocialCollectionHandler so that it no longer creates a collection on the GSA. It now only validates the collection name and logs a warning that the collection must be created manually.
The Connector for SharePoint User Profiles 4.0 (aka adaptor)
intentionally requires the collection to be created manually, so this
change is forward-compatible if not backward-compatible.

This is part of removing the GSA Admin API so that we can upgrade
Guava. The GSA Admin API depends on the GData library, which in turn
depends on Guava version 10 or earlier.

Code review: https://codereview.appspot.com/118930043/