Rollback r3312 to upgrade Guava from 10.0.1 to 11.0.2.
The file system, Notes, and SharePoint connectors have been updated to
remove the GSA Admin API. The GSA Admin API depends on the GData library,
which in turn depends on Guava version 10 or earlier.

This change requires Notes connector trunk@417, which will not be on
the 3.2.x release branch. The file system and SharePoint connectors
are compatible with this change, because they never directly used the
conflicting GData client.

Code review: https://codereview.appspot.com/189450043/